server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Нурсиль <n.sultanoff.2002s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.1"}
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.216.0"}
bcrypt = "^4.1.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
mock = "^5.1.0"
pytest-mock = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
asyncio_mode="auto"